refactor: move operation to feature folder
