+++ packages/api-reference/src/features/Operation/ParameterListItem.vue
@@ -1,17 +1,17 @@
 <script setup lang="ts">
+import { SchemaProperty } from '@/components/Content/Schema'
+import ScreenReader from '@/components/ScreenReader.vue'
 import { Disclosure, DisclosureButton, DisclosurePanel } from '@headlessui/vue'
 import { ScalarIcon } from '@scalar/components'
-import type { ContentType, Parameters } from '@scalar/types/legacy'
+import type { ContentType, Parameter } from '@scalar/types/legacy'
 import { computed, ref } from 'vue'
 
-import ScreenReader from '../../ScreenReader.vue'
-import { SchemaProperty } from '../Schema'
 import ContentTypeSelect from './ContentTypeSelect.vue'
 import ParameterHeaders from './ParameterHeaders.vue'
 
 const props = withDefaults(
   defineProps<{
-    parameter: Parameters
+    parameter: Parameter
     showChildren?: boolean
     collapsableItems?: boolean
   }>(),
