+++ packages/api-reference/src/features/Operation/RequestBody.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
+import { Schema } from '@/components/Content/Schema'
 import { ScalarMarkdown } from '@scalar/components'
 import type { ContentType, RequestBody } from '@scalar/types/legacy'
 import { computed, ref } from 'vue'
-
-import { Schema } from '../Schema'
 
 const prop = defineProps<{ requestBody?: RequestBody }>()
