+++ packages/api-reference/src/features/Operation/layouts/ClassicLayout.vue
@@ -1,15 +1,15 @@
 <script setup lang="ts">
+import { Anchor } from '@/components/Anchor'
+import { HttpMethod } from '@/components/HttpMethod'
+import { SectionAccordion } from '@/components/Section'
+import { ExampleRequest } from '@/features/ExampleRequest'
+import { ExampleResponses } from '@/features/ExampleResponses'
+import { HIDE_TEST_REQUEST_BUTTON_SYMBOL } from '@/helpers'
 import { ScalarIcon, ScalarIconButton } from '@scalar/components'
 import type { TransformedOperation } from '@scalar/types/legacy'
 import { useClipboard } from '@scalar/use-hooks/useClipboard'
 import { inject } from 'vue'
 
-import { ExampleRequest } from '../../../../features/ExampleRequest'
-import { ExampleResponses } from '../../../../features/ExampleResponses'
-import { HIDE_TEST_REQUEST_BUTTON_SYMBOL } from '../../../../helpers'
-import { Anchor } from '../../../Anchor'
-import { HttpMethod } from '../../../HttpMethod'
-import { SectionAccordion } from '../../../Section'
 import OperationDescription from '../OperationDescription.vue'
 import OperationParameters from '../OperationParameters.vue'
 import OperationPath from '../OperationPath.vue'
