+++ packages/api-reference/vite.config.ts
@@ -1,12 +1,19 @@
 import { autoCSSInject, createViteBuildOptions } from '@scalar/build-tooling'
 import vue from '@vitejs/plugin-vue'
+import { URL, fileURLToPath } from 'node:url'
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
+  plugins: [vue()],
   define: {
     'process.env.NODE_ENV': '"production"',
   },
-  plugins: [vue()],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+    },
+    dedupe: ['vue'],
+  },
   build: createViteBuildOptions({
     entry: ['src/index.ts'],
     options: {
@@ -17,9 +24,6 @@ export default defineConfig({
       },
     },
   }),
-  resolve: {
-    dedupe: ['vue'],
-  },
   test: {
     coverage: {
       enabled: true,
